TrackFitting: hide KalmanFitterError implementation
