+++ Core/CMakeLists.txt
@@ -51,6 +51,7 @@ add_subdirectory(src/Material)
 add_subdirectory(src/Propagator)
 add_subdirectory(src/Surfaces)
 add_subdirectory(src/TrackFinding)
+add_subdirectory(src/TrackFitting)
 add_subdirectory(src/Utilities)
 add_subdirectory(src/Vertexing)
 add_subdirectory(src/Visualization)
