+++ Core/src/TrackFitting/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources_local(
+  ActsCore
+  PRIVATE
+    KalmanFitterError.cpp
+)
